ADD remaining files required for implementation
Added in index.html, script.js and styles.css. Boilerplate and general
layout added to index.html. Base styles, variables and a little styling
added to styles.css. Currently script.js is empty but will contain the
logic for populating the main section with a dynamic title, description
and links according to which tab is clicked.